backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1

Flask-Marshmallow==0.15.0
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0

PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography==42.0.8
